fn-impl: reorganize files
